Move RasterFile to new module
